fix: wrap canvas page useSearchParams in Suspense boundary
useSearchParams() in a client component without a Suspense boundary
causes Next.js 16 to fail during static page generation, breaking
Docker builds. Extract the client logic into a separate component
and wrap it with <Suspense> in the page.

Closes #70
